# Use awk instead of cat&echo to boost json concating performance