--- a/generate.sh
+++ b/generate.sh
@@ -110,7 +110,8 @@ fi
   $("${QUERY_CMD[@]}") > /dev/null
 
 echo "[" > "${COMPDB_FILE}"
-find "${EXEC_ROOT}" -name '*.compile_commands.json' -not -empty -exec bash -c 'cat "$1" && echo ,' _ {} \; \
+find "${EXEC_ROOT}" -name '*.compile_commands.json' -not -empty \
+    | awk '{fn=$0; RS="^$"; getline s<fn; close(fn); RS="\n"; print s "," }' \
   >> "${COMPDB_FILE}"
 echo "]" >> "${COMPDB_FILE}"