@gou4shi1 2021-05-08
Nice job! It save me 20s! Why the simple cat&echo is such slow?