@siddharthab 2021-09-08
Thank you for your PR and my apologies for lack of attention to this repo. I think the echo command is a shell builtin and can be very slow.